* Sanitize some inputs and handle errors better * Refactor UI code into multiple modules * Add lots of tests * Add Rust coding guidelines for copilot * Resolve several clippy warnings
