+++ data-gov/tools/cli/ui/mod.rs
@@ -1,582 +1,104 @@
-mod colors;
+pub mod colors;
 
-use self::colors::{ColorHelper, ColorMode};
+mod commands;
+mod display;
+mod handlers;
+mod repl;
+mod reporter;
+
 use clap::{Arg, ArgMatches, Command};
-use indicatif::{ProgressBar, ProgressStyle};
-use is_terminal::IsTerminal;
-use rustyline::{DefaultEditor, Result as RustyResult};
-use tokio::runtime::Runtime;
-
-use std::collections::HashMap;
-use std::io;
 use std::path::PathBuf;
 use std::str::FromStr;
-use std::sync::OnceLock;
-use std::sync::{Arc, Mutex};
+use std::sync::{Arc, OnceLock};
+use tokio::runtime::Runtime;
 
-use data_gov::ui::{
-    DownloadBatch, DownloadFailed, DownloadFinished, DownloadProgress, DownloadStarted,
-    StatusReporter,
-};
-use data_gov::{DataGovClient, DataGovConfig, OperatingMode, ckan::models::Package};
+use self::colors::{ColorHelper, ColorMode};
+use self::commands::ReplCommand;
+use self::handlers::execute_command;
+use self::repl::DataGovRepl;
+use self::reporter::CliStatusReporter;
+
+use data_gov::{DataGovClient, DataGovConfig, OperatingMode};
 
 /// Global color helper - will be set at startup
 static COLOR_HELPER: OnceLock<ColorHelper> = OnceLock::new();
 
 /// Helper functions for color formatting
-fn color_red_bold(text: &str) -> String {
+pub fn color_red_bold(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.style().red(text).bold().to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_green_bold(text: &str) -> String {
+pub fn color_green_bold(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.style().green(text).bold().to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_blue_bold(text: &str) -> String {
+pub fn color_blue_bold(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.style().blue(text).bold().to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_yellow_bold(text: &str) -> String {
+pub fn color_yellow_bold(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.style().yellow(text).bold().to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_cyan(text: &str) -> String {
+pub fn color_cyan(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.cyan(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_blue(text: &str) -> String {
+pub fn color_blue(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.blue(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_green(text: &str) -> String {
+pub fn color_green(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.green(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_yellow(text: &str) -> String {
+pub fn color_yellow(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.yellow(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_red(text: &str) -> String {
+pub fn color_red(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.red(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_dimmed(text: &str) -> String {
+pub fn color_dimmed(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.dimmed(text).to_string())
         .unwrap_or_else(|| text.to_string())
 }
 
-fn color_bold(text: &str) -> String {
+pub fn color_bold(text: &str) -> String {
     COLOR_HELPER
         .get()
         .map(|h| h.bold(text).to_string())
         .unwrap_or_else(|| text.to_string())
-}
-
-struct CliStatusReporter {
-    color_helper: ColorHelper,
-    show_progress: bool,
-    fancy_progress: bool,
-    progress: Mutex<HashMap<String, ProgressBar>>,
-    print_lock: Mutex<()>,
-}
-
-impl CliStatusReporter {
-    fn new(color_helper: ColorHelper) -> Self {
-        let show_progress = std::env::var("NO_PROGRESS").is_err();
-        let fancy_progress = show_progress
-            && std::io::stdout().is_terminal()
-            && std::env::var("FORCE_SIMPLE_PROGRESS").is_err();
-
-        Self {
-            color_helper,
-            show_progress,
-            fancy_progress,
-            progress: Mutex::new(HashMap::new()),
-            print_lock: Mutex::new(()),
-        }
-    }
-
-    fn display_name(&self, resource_name: &Option<String>, output_path: &PathBuf) -> String {
-        if let Some(name) = resource_name {
-            return name.clone();
-        }
-
-        output_path
-            .file_name()
-            .and_then(|name| name.to_str())
-            .map(|s| s.to_string())
-            .unwrap_or_else(|| "file".to_string())
-    }
-
-    fn progress_key(path: &PathBuf) -> String {
-        path.to_string_lossy().into_owned()
-    }
-}
-
-impl StatusReporter for CliStatusReporter {
-    fn on_download_batch(&self, event: &DownloadBatch) {
-        if !self.show_progress {
-            return;
-        }
-
-        let _guard = self.print_lock.lock().unwrap();
-        match &event.dataset_name {
-            Some(dataset) => println!(
-                "{} {} resources for '{}'...",
-                color_cyan("Downloading"),
-                event.resource_count,
-                dataset
-            ),
-            None => println!(
-                "{} {} resources...",
-                color_cyan("Downloading"),
-                event.resource_count
-            ),
-        }
-    }
-
-    fn on_download_started(&self, event: &DownloadStarted) {
-        if !self.show_progress {
-            return;
-        }
-
-        let name = self.display_name(&event.resource_name, &event.output_path);
-
-        if self.fancy_progress {
-            let pb = match event.total_bytes {
-                Some(total) => ProgressBar::new(total),
-                None => ProgressBar::new_spinner(),
-            };
-
-            let template = if event.total_bytes.is_some() {
-                "{msg} [{bar:40.cyan/blue}] {bytes}/{total_bytes} ({bytes_per_sec}, {eta})"
-            } else {
-                "{msg} [{spinner:.cyan/blue}] {bytes} ({bytes_per_sec})"
-            };
-
-            pb.set_style(
-                ProgressStyle::default_bar()
-                    .template(template)
-                    .unwrap_or_else(|_| ProgressStyle::default_bar())
-                    .progress_chars("█▉▊▋▌▍▎▏ "),
-            );
-            pb.set_message(format!("Downloading {}", name));
-
-            let key = Self::progress_key(&event.output_path);
-            self.progress.lock().unwrap().insert(key, pb);
-        } else {
-            let _guard = self.print_lock.lock().unwrap();
-            if let Some(total) = event.total_bytes {
-                println!("Downloading {} ({} bytes)...", name, total);
-            } else {
-                println!("Downloading {} ...", name);
-            }
-        }
-    }
-
-    fn on_download_progress(&self, event: &DownloadProgress) {
-        if !self.fancy_progress {
-            return;
-        }
-
-        let key = Self::progress_key(&event.output_path);
-        if let Some(pb) = self.progress.lock().unwrap().get(&key) {
-            if let Some(total) = event.total_bytes {
-                pb.set_length(total);
-            }
-            pb.set_position(event.downloaded_bytes);
-        }
-    }
-
-    fn on_download_finished(&self, event: &DownloadFinished) {
-        let name = self.display_name(&event.resource_name, &event.output_path);
-
-        if self.fancy_progress {
-            let key = Self::progress_key(&event.output_path);
-            if let Some(pb) = self.progress.lock().unwrap().remove(&key) {
-                pb.finish_with_message(format!("Downloaded {}", name));
-                return;
-            }
-        }
-
-        if self.show_progress {
-            let _guard = self.print_lock.lock().unwrap();
-            println!("{} {}", self.color_helper.green("✓ Downloaded"), name);
-        }
-    }
-
-    fn on_download_failed(&self, event: &DownloadFailed) {
-        let name = event
-            .output_path
-            .as_ref()
-            .map(Self::progress_key)
-            .unwrap_or_else(|| {
-                event
-                    .resource_name
-                    .clone()
-                    .unwrap_or_else(|| "file".to_string())
-            });
-
-        if self.fancy_progress {
-            if let Some(path) = &event.output_path {
-                let key = Self::progress_key(path);
-                if let Some(pb) = self.progress.lock().unwrap().remove(&key) {
-                    pb.abandon_with_message(format!("Failed {}: {}", name, event.error));
-                    return;
-                }
-            }
-        }
-
-        let _guard = self.print_lock.lock().unwrap();
-        println!("{} {} ({})", color_red_bold("Failed:"), name, event.error);
-    }
-}
-
-/// REPL Commands
-#[derive(Debug, Clone)]
-enum ReplCommand {
-    Search {
-        query: String,
-        limit: Option<i32>,
-    },
-    Show {
-        dataset_id: String,
-    },
-    Download {
-        dataset_id: String,
-        resource_selector: Option<ResourceSelector>,
-    },
-    List {
-        what: String,
-    }, // organizations, formats, etc.
-    SetDir {
-        path: PathBuf,
-    },
-    Info,
-    Help,
-    Quit,
-}
-
-/// Resource selector for download command
-#[derive(Debug, Clone)]
-enum ResourceSelector {
-    Index(usize),
-    Name(String),
-}
-
-/// Parse a command string respecting quoted arguments
-/// Example: `foo bar "baz qux"` -> ["foo", "bar", "baz qux"]
-fn parse_command_args(s: &str) -> Vec<String> {
-    let mut args = Vec::new();
-    let mut current_arg = String::new();
-    let mut in_quotes = false;
-    let mut chars = s.trim().chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        match ch {
-            '"' => {
-                in_quotes = !in_quotes;
-            }
-            ' ' | '\t' if !in_quotes => {
-                if !current_arg.is_empty() {
-                    args.push(current_arg.clone());
-                    current_arg.clear();
-                }
-            }
-            _ => {
-                current_arg.push(ch);
-            }
-        }
-    }
-
-    if !current_arg.is_empty() {
-        args.push(current_arg);
-    }
-
-    args
-}
-
-impl FromStr for ReplCommand {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let parts = parse_command_args(s);
-
-        if parts.is_empty() {
-            return Err("Empty command".to_string());
-        }
-
-        match parts[0].to_lowercase().as_str() {
-            "search" | "s" => {
-                if parts.len() < 2 {
-                    return Err("Usage: search <query> [limit]".to_string());
-                }
-                let query = parts[1..].join(" ");
-                let limit = if parts.len() > 2 {
-                    parts.last().unwrap().parse().ok()
-                } else {
-                    None
-                };
-                Ok(ReplCommand::Search { query, limit })
-            }
-            "show" | "describe" | "d" => {
-                if parts.len() != 2 {
-                    return Err("Usage: show <dataset_id>".to_string());
-                }
-                Ok(ReplCommand::Show {
-                    dataset_id: parts[1].clone(),
-                })
-            }
-            "download" | "dl" => {
-                if parts.len() < 2 || parts.len() > 3 {
-                    return Err("Usage: download <dataset_id> [resource_index_or_name]".to_string());
-                }
-                let resource_selector = if parts.len() == 3 {
-                    // Try to parse as index first, otherwise treat as name
-                    if let Ok(index) = parts[2].parse::<usize>() {
-                        Some(ResourceSelector::Index(index))
-                    } else {
-                        Some(ResourceSelector::Name(parts[2].clone()))
-                    }
-                } else {
-                    None
-                };
-                Ok(ReplCommand::Download {
-                    dataset_id: parts[1].clone(),
-                    resource_selector,
-                })
-            }
-            "list" | "ls" => {
-                if parts.len() != 2 {
-                    return Err("Usage: list <organizations|orgs>".to_string());
-                }
-                Ok(ReplCommand::List {
-                    what: parts[1].clone(),
-                })
-            }
-            "setdir" | "cd" => {
-                if parts.len() != 2 {
-                    return Err("Usage: setdir <path>".to_string());
-                }
-                Ok(ReplCommand::SetDir {
-                    path: PathBuf::from(&parts[1]),
-                })
-            }
-            "info" | "status" => Ok(ReplCommand::Info),
-            "help" | "h" | "?" => Ok(ReplCommand::Help),
-            "quit" | "exit" | "q" => Ok(ReplCommand::Quit),
-            _ => Err(format!("Unknown command: {}", parts[0])),
-        }
-    }
-}
-
-/// REPL state and logic
-struct DataGovRepl {
-    client: DataGovClient,
-    rt: Runtime,
-}
-
-impl DataGovRepl {
-    fn new(client: DataGovClient) -> io::Result<Self> {
-        let rt = Runtime::new()?;
-        Ok(Self { client, rt })
-    }
-
-    fn run(&mut self) -> RustyResult<()> {
-        println!("{}", color_blue_bold("🇺🇸 Data.gov Interactive Explorer"));
-        println!(
-            "{}",
-            color_dimmed("Type 'help' for available commands, 'quit' to exit")
-        );
-        println!();
-
-        let mut rl = DefaultEditor::new()?;
-
-        loop {
-            let readline = rl.readline(&format!("{} ", color_green_bold("data.gov>")));
-
-            match readline {
-                Ok(line) => {
-                    let trimmed = line.trim();
-
-                    // Skip empty lines and comments
-                    if trimmed.is_empty() || trimmed.starts_with('#') {
-                        continue;
-                    }
-
-                    rl.add_history_entry(line.as_str())?;
-
-                    match ReplCommand::from_str(&line) {
-                        Ok(command) => {
-                            if let ReplCommand::Quit = command {
-                                println!("Goodbye! 👋");
-                                break;
-                            }
-
-                            if let Err(e) = self.handle_command(command) {
-                                println!("{} {}", color_red_bold("Error:"), e);
-                            }
-                        }
-                        Err(e) => {
-                            println!("{} {}", color_red_bold("Invalid command:"), e);
-                        }
-                    }
-                }
-                Err(rustyline::error::ReadlineError::Interrupted) => {
-                    println!("CTRL-C");
-                    break;
-                }
-                Err(rustyline::error::ReadlineError::Eof) => {
-                    println!("CTRL-D");
-                    break;
-                }
-                Err(err) => {
-                    println!("Error: {:?}", err);
-                    break;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn handle_command(&mut self, command: ReplCommand) -> Result<(), Box<dyn std::error::Error>> {
-        // Handle REPL-specific commands
-        match &command {
-            ReplCommand::SetDir { path } => {
-                // Create new config with updated directory
-                let new_config = DataGovConfig::new().with_download_dir(path.clone());
-
-                // Validate directory
-                self.rt.block_on(async {
-                    let temp_client = DataGovClient::with_config(new_config.clone())?;
-                    temp_client.validate_download_dir().await?;
-                    self.client = temp_client;
-                    Ok::<(), data_gov::DataGovError>(())
-                })?;
-
-                println!(
-                    "{} Download directory set to: {}",
-                    color_green_bold("Success!"),
-                    color_blue(&path.display().to_string())
-                );
-                return Ok(());
-            }
-            _ => {}
-        }
-
-        // Handle Help command specially for REPL mode
-        match &command {
-            ReplCommand::Help => {
-                self.print_repl_help();
-                return Ok(());
-            }
-            _ => {}
-        }
-
-        // Use shared command execution logic for other commands
-        execute_command(&self.client, &self.rt, command)?;
-        Ok(())
-    }
-
-    fn print_repl_help(&self) {
-        println!("\n{}", color_blue_bold("📚 Available Commands"));
-        println!();
-
-        let commands = vec![
-            (
-                "search <query> [limit]",
-                "Search for datasets",
-                "search climate data 20",
-            ),
-            (
-                "show <dataset_id>",
-                "Show detailed dataset information",
-                "show consumer-complaint-database",
-            ),
-            (
-                "download <dataset_id> [index|name]",
-                "Download by index or name (partial match)",
-                "download my-dataset 0 | download my-dataset csv",
-            ),
-            (
-                "list organizations",
-                "List government organizations",
-                "list orgs",
-            ),
-            (
-                "setdir <path>",
-                "Set base download directory",
-                "setdir ./downloads",
-            ),
-            ("info", "Show session information", "info"),
-            ("help", "Show this help message", "help"),
-            ("quit", "Exit the REPL", "quit"),
-        ];
-
-        for (cmd, desc, example) in commands {
-            println!("{:25} {}", color_green_bold(cmd), desc);
-            println!(
-                "{:25} {}: {}",
-                "",
-                color_dimmed("Example"),
-                color_blue(example)
-            );
-            println!();
-        }
-
-        println!("{}", color_yellow_bold("💡 Pro Tips:"));
-        println!(
-            "  • Use short commands: {} for search, {} for show, {} for download",
-            color_green("s"),
-            color_green("d"),
-            color_green("dl")
-        );
-        println!(
-            "  • Search supports multiple words: {}",
-            color_blue("search energy solar wind")
-        );
-        println!("  • Downloads are organized by dataset name in subdirectories");
-        println!("  • Download without index downloads all resources");
-        println!(
-            "  • Create scripts with {} for automation",
-            color_blue("#!/usr/bin/env data-gov")
-        );
-        println!(
-            "  • Run {} to use CLI mode",
-            color_blue("data-gov search \"climate\" 10")
-        );
-        println!();
-    }
 }
 
 pub fn run() -> Result<(), Box<dyn std::error::Error>> {
@@ -732,750 +254,4 @@ fn run_cli_mode(
     }
 
     Ok(())
-}
-
-/// Execute a command (shared between REPL and CLI modes)
-fn execute_command(
-    client: &DataGovClient,
-    rt: &Runtime,
-    command: ReplCommand,
-) -> Result<(), Box<dyn std::error::Error>> {
-    match command {
-        ReplCommand::Search { query, limit } => {
-            println!("{} '{}'...", color_cyan("Searching for"), query);
-
-            let results = rt.block_on(client.search(&query, limit, None, None, None))?;
-
-            println!(
-                "\n{} {} results:\n",
-                color_green_bold("Found"),
-                results.count.unwrap_or(0)
-            );
-
-            if let Some(packages) = results.results {
-                for (i, package) in packages.iter().enumerate().take(20) {
-                    println!(
-                        "{}. {} {}",
-                        color_blue_bold(&format!("{:2}", i + 1)),
-                        color_yellow_bold(&package.name),
-                        color_dimmed(package.title.as_deref().unwrap_or(""))
-                    );
-
-                    if let Some(notes) = &package.notes {
-                        let truncated = if notes.len() > 100 {
-                            format!("{}...", &notes[..100])
-                        } else {
-                            notes.clone()
-                        };
-                        println!("   {}", color_dimmed(&truncated));
-                    }
-                    println!();
-                }
-
-                if packages.len() > 20 {
-                    println!("... and {} more results", packages.len() - 20);
-                }
-            }
-        }
-
-        ReplCommand::Show { dataset_id } => {
-            println!("{} dataset '{}'...", color_cyan("Fetching"), dataset_id);
-
-            let package = rt.block_on(client.get_dataset(&dataset_id))?;
-            print_package_details(&package);
-        }
-
-        ReplCommand::Download {
-            dataset_id,
-            resource_selector,
-        } => {
-            println!("{} dataset '{}'...", color_cyan("Fetching"), dataset_id);
-
-            let package = rt.block_on(client.get_dataset(&dataset_id))?;
-            let resources = DataGovClient::get_downloadable_resources(&package);
-
-            if resources.is_empty() {
-                println!(
-                    "{} No downloadable resources found in this dataset.",
-                    color_yellow_bold("Warning:")
-                );
-                return Ok(());
-            }
-
-            match resource_selector {
-                Some(ResourceSelector::Index(index)) => {
-                    // Download by index
-                    if index >= resources.len() {
-                        println!(
-                            "{} Resource index {} is out of range (0-{})",
-                            color_red_bold("Error:"),
-                            index,
-                            resources.len() - 1
-                        );
-                        return Ok(());
-                    }
-
-                    let resource = &resources[index];
-                    println!("{} resource {}...", color_cyan("Downloading"), index);
-
-                    // Download to dataset-specific directory
-                    let dataset_dir = client.download_dir().join(&dataset_id);
-                    let path = rt.block_on(client.download_resource(resource, Some(&dataset_dir)))?;
-
-                    println!(
-                        "{} Downloaded to: {}",
-                        color_green_bold("Success!"),
-                        color_blue(&path.display().to_string())
-                    );
-                }
-                Some(ResourceSelector::Name(name)) => {
-                    // Download by name (case-insensitive, may match multiple)
-                    let name_lower = name.to_lowercase();
-                    let matching_resources: Vec<(usize, &data_gov::ckan::models::Resource)> = resources
-                        .iter()
-                        .enumerate()
-                        .filter(|(_, r)| {
-                            r.name
-                                .as_ref()
-                                .map(|n| n.to_lowercase().contains(&name_lower))
-                                .unwrap_or(false)
-                        })
-                        .collect();
-
-                    if matching_resources.is_empty() {
-                        println!(
-                            "{} No resources found matching name '{}'",
-                            color_yellow_bold("Warning:"),
-                            name
-                        );
-                        println!("\nAvailable resources:");
-                        for (i, r) in resources.iter().enumerate() {
-                            let rname = r.name.as_deref().unwrap_or("(unnamed)");
-                            let format = r.format.as_deref().unwrap_or("?");
-                            println!("  {} {} [{}]", i, rname, format);
-                        }
-                        return Ok(());
-                    }
-
-                    if matching_resources.len() == 1 {
-                        let (idx, resource) = matching_resources[0];
-                        println!(
-                            "{} resource {} matching '{}'...",
-                            color_cyan("Downloading"),
-                            idx,
-                            name
-                        );
-
-                        let dataset_dir = client.download_dir().join(&dataset_id);
-                        let path = rt.block_on(client.download_resource(resource, Some(&dataset_dir)))?;
-
-                        println!(
-                            "{} Downloaded to: {}",
-                            color_green_bold("Success!"),
-                            color_blue(&path.display().to_string())
-                        );
-                    } else {
-                        // Multiple matches - download all
-                        println!(
-                            "{} {} resources matching '{}'...",
-                            color_cyan("Downloading"),
-                            matching_resources.len(),
-                            name
-                        );
-
-                        let dataset_dir = client.download_dir().join(&dataset_id);
-                        let resources_to_download: Vec<_> = matching_resources.iter().map(|(_, r)| (*r).clone()).collect();
-                        let results = rt.block_on(client.download_resources(&resources_to_download, Some(&dataset_dir)));
-
-                        let mut success_count = 0;
-                        let mut error_count = 0;
-
-                        for (result, (idx, _)) in results.iter().zip(matching_resources.iter()) {
-                            match result {
-                                Ok(path) => {
-                                    success_count += 1;
-                                    println!(
-                                        "  {} Resource {}: {}",
-                                        color_green("✓"),
-                                        idx,
-                                        color_blue(&path.display().to_string())
-                                    );
-                                }
-                                Err(e) => {
-                                    error_count += 1;
-                                    println!(
-                                        "  {} Resource {}: {}",
-                                        color_red("✗"),
-                                        idx,
-                                        color_red(&e.to_string())
-                                    );
-                                }
-                            }
-                        }
-
-                        println!(
-                            "\n{} {} downloaded, {} errors",
-                            color_bold("Summary:"),
-                            color_green(&success_count.to_string()),
-                            color_red(&error_count.to_string())
-                        );
-                    }
-                }
-                None => {
-                    // Download all resources to dataset-specific directory
-                    let dataset_dir = client.download_dir().join(&dataset_id);
-                    let results = rt.block_on(client.download_resources(&resources, Some(&dataset_dir)));
-
-                    let mut success_count = 0;
-                    let mut error_count = 0;
-
-                    for (i, result) in results.iter().enumerate() {
-                        match result {
-                            Ok(path) => {
-                                success_count += 1;
-                                println!(
-                                    "  {} Resource {}: {}",
-                                    color_green("✓"),
-                                    i,
-                                    color_blue(&path.display().to_string())
-                                );
-                            }
-                            Err(e) => {
-                                error_count += 1;
-                                println!(
-                                    "  {} Resource {}: {}",
-                                    color_red("✗"),
-                                    i,
-                                    color_red(&e.to_string())
-                                );
-                            }
-                        }
-                    }
-
-                    println!(
-                        "\n{} {} downloaded, {} errors",
-                        color_bold("Summary:"),
-                        color_green(&success_count.to_string()),
-                        color_red(&error_count.to_string())
-                    );
-                }
-            }
-        }
-
-        ReplCommand::List { what } => match what.to_lowercase().as_str() {
-            "organizations" | "orgs" => {
-                println!("{} organizations...", color_cyan("Fetching"));
-
-                let orgs = rt.block_on(client.list_organizations(Some(50)))?;
-
-                println!("\n{} organizations:", color_green_bold("Government"));
-                for (i, org) in orgs.iter().enumerate() {
-                    println!(
-                        "{}. {}",
-                        color_blue_bold(&format!("{:2}", i + 1)),
-                        color_yellow(org)
-                    );
-                }
-            }
-            _ => {
-                println!("{} Unknown list type: {}", color_red_bold("Error:"), what);
-                println!("Available: {}", color_blue("organizations"));
-            }
-        },
-
-        ReplCommand::Info => {
-            println!("\n{}", color_blue_bold("📊 Client Information"));
-            println!(
-                "Download directory: {}",
-                color_blue(&client.download_dir().display().to_string())
-            );
-            println!(
-                "CKAN endpoint: {}",
-                color_blue("https://catalog.data.gov/api/3")
-            );
-        }
-
-        ReplCommand::SetDir { .. } => {
-            println!(
-                "{} SetDir command is only available in interactive REPL mode",
-                color_red_bold("Error:")
-            );
-        }
-
-        ReplCommand::Help => {
-            print_cli_help();
-        }
-
-        ReplCommand::Quit => {
-            // Not applicable in CLI mode
-        }
-    }
-
-    Ok(())
-}
-
-/// Print help for CLI mode
-fn print_cli_help() {
-    println!("\n{}", color_blue_bold("📚 CLI Mode Commands"));
-    println!();
-
-    let commands = vec![
-        (
-            "search <query> [limit]",
-            "Search for datasets",
-            "search \"climate data\" 20",
-        ),
-        (
-            "show <dataset_id>",
-            "Show detailed dataset information",
-            "show consumer-complaint-database",
-        ),
-        (
-            "download <dataset_id> [index|name]",
-            "Download by index or name (partial match)",
-            "download my-dataset csv",
-        ),
-        (
-            "list organizations",
-            "List government organizations",
-            "list organizations",
-        ),
-        ("info", "Show client information", "info"),
-    ];
-
-    for (cmd, desc, example) in commands {
-        println!("{:30} {}", color_green_bold(cmd), desc);
-        println!(
-            "{:30} {}: data-gov {}",
-            "",
-            color_dimmed("Example"),
-            color_blue(example)
-        );
-        println!();
-    }
-
-    println!("{}", color_yellow_bold("💡 Interactive Mode:"));
-    println!(
-        "  Run without arguments to start interactive REPL: {}",
-        color_blue("data-gov")
-    );
-    println!();
-}
-
-/// Print package details (shared between REPL and CLI modes)
-fn print_package_details(package: &Package) {
-    println!("\n{}", color_blue_bold("📦 Dataset Details"));
-    println!("{}: {}", color_bold("Name"), color_yellow(&package.name));
-
-    if let Some(title) = &package.title {
-        println!("{}: {}", color_bold("Title"), title);
-    }
-
-    if let Some(notes) = &package.notes {
-        println!("\n{}: ", color_bold("Description"));
-        println!("{}", color_dimmed(notes));
-    }
-
-    if let Some(license_title) = &package.license_title {
-        println!(
-            "\n{}: {}",
-            color_bold("License"),
-            color_green(license_title)
-        );
-    }
-
-    if let Some(author) = &package.author {
-        println!("{}: {}", color_bold("Author"), author);
-    }
-
-    if let Some(maintainer) = &package.maintainer {
-        println!("{}: {}", color_bold("Maintainer"), maintainer);
-    }
-
-    // Display resources
-    let resources = DataGovClient::get_downloadable_resources(package);
-    if !resources.is_empty() {
-        println!(
-            "\n{} {} downloadable resources:",
-            color_bold("📁"),
-            resources.len()
-        );
-
-        for (i, resource) in resources.iter().enumerate() {
-            let name = resource.name.as_deref().unwrap_or("Unnamed");
-            let format = resource.format.as_deref().unwrap_or("Unknown");
-            let size = resource
-                .size
-                .map(|s| format!(" ({})", s))
-                .unwrap_or_default();
-
-            println!(
-                "  {}. {} {} {}{}",
-                color_blue_bold(&i.to_string()),
-                color_yellow(name),
-                color_green(&format!("[{}]", format)),
-                color_dimmed(&size),
-                ""
-            );
-
-            if let Some(desc) = &resource.description {
-                if !desc.is_empty() {
-                    let truncated = if desc.len() > 80 {
-                        format!("{}...", &desc[..80])
-                    } else {
-                        desc.clone()
-                    };
-                    println!("     {}", color_dimmed(&truncated));
-                }
-            }
-        }
-
-        println!(
-            "\n{} Use 'data-gov download {}' to download all resources",
-            color_bold("💡"),
-            color_yellow(&package.name)
-        );
-        println!(
-            "{} Use 'data-gov download {} <index|name>' to download by index or name",
-            color_bold("💡"),
-            color_yellow(&package.name)
-        );
-    } else {
-        println!("\n{} No downloadable resources found", color_yellow("⚠️"));
-    }
-
-    println!();
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_download_command_with_index() {
-        let result = ReplCommand::from_str("download my-dataset 0");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            assert!(matches!(resource_selector, Some(ResourceSelector::Index(0))));
-        } else {
-            panic!("Expected Download command with Index(0)");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_with_large_index() {
-        let result = ReplCommand::from_str("download my-dataset 42");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            assert!(matches!(resource_selector, Some(ResourceSelector::Index(42))));
-        } else {
-            panic!("Expected Download command with Index(42)");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_with_name() {
-        let result = ReplCommand::from_str("download my-dataset csv");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "csv");
-            } else {
-                panic!("Expected Download command with Name(csv)");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_with_partial_name() {
-        let result = ReplCommand::from_str("download dataset-id complaints");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "dataset-id");
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "complaints");
-            } else {
-                panic!("Expected Download command with Name(complaints)");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_no_selector() {
-        let result = ReplCommand::from_str("download my-dataset");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            assert!(resource_selector.is_none());
-        } else {
-            panic!("Expected Download command with no selector");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_dl_alias() {
-        let result = ReplCommand::from_str("dl my-dataset 0");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            assert!(matches!(resource_selector, Some(ResourceSelector::Index(0))));
-        } else {
-            panic!("Expected Download command using 'dl' alias");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_with_extension() {
-        // This should be treated as a name since it's not a valid number
-        let result = ReplCommand::from_str("download my-dataset data.csv");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "data.csv");
-            } else {
-                panic!("Expected Download command with Name(data.csv)");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_command_too_many_args() {
-        let result = ReplCommand::from_str("download dataset-id resource extra-arg");
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Usage: download <dataset_id> [resource_index_or_name]"));
-    }
-
-    #[test]
-    fn test_parse_download_command_too_few_args() {
-        let result = ReplCommand::from_str("download");
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Usage: download <dataset_id> [resource_index_or_name]"));
-    }
-
-    #[test]
-    fn test_resource_name_matching_case_insensitive() {
-        // Create test resources
-        let resources = vec![
-            data_gov::ckan::models::Resource {
-                name: Some("Data.CSV".to_string()),
-                format: Some("CSV".to_string()),
-                url: Some("https://example.com/data.csv".to_string()),
-                ..Default::default()
-            },
-            data_gov::ckan::models::Resource {
-                name: Some("report.json".to_string()),
-                format: Some("JSON".to_string()),
-                url: Some("https://example.com/report.json".to_string()),
-                ..Default::default()
-            },
-            data_gov::ckan::models::Resource {
-                name: Some("ARCHIVE.CSV".to_string()),
-                format: Some("CSV".to_string()),
-                url: Some("https://example.com/archive.csv".to_string()),
-                ..Default::default()
-            },
-        ];
-
-        // Test case-insensitive matching for "csv"
-        let name_lower = "csv".to_lowercase();
-        let matching: Vec<_> = resources
-            .iter()
-            .enumerate()
-            .filter(|(_, r)| {
-                r.name
-                    .as_ref()
-                    .map(|n| n.to_lowercase().contains(&name_lower))
-                    .unwrap_or(false)
-            })
-            .collect();
-
-        assert_eq!(matching.len(), 2);
-        assert_eq!(matching[0].0, 0); // Data.CSV
-        assert_eq!(matching[1].0, 2); // ARCHIVE.CSV
-    }
-
-    #[test]
-    fn test_resource_name_matching_partial() {
-        // Create test resources
-        let resources = vec![
-            data_gov::ckan::models::Resource {
-                name: Some("complaints-2023.csv".to_string()),
-                format: Some("CSV".to_string()),
-                url: Some("https://example.com/file1.csv".to_string()),
-                ..Default::default()
-            },
-            data_gov::ckan::models::Resource {
-                name: Some("data.json".to_string()),
-                format: Some("JSON".to_string()),
-                url: Some("https://example.com/data.json".to_string()),
-                ..Default::default()
-            },
-            data_gov::ckan::models::Resource {
-                name: Some("complaints-2024.csv".to_string()),
-                format: Some("CSV".to_string()),
-                url: Some("https://example.com/file2.csv".to_string()),
-                ..Default::default()
-            },
-        ];
-
-        // Test partial matching for "complaint"
-        let name_lower = "complaint".to_lowercase();
-        let matching: Vec<_> = resources
-            .iter()
-            .enumerate()
-            .filter(|(_, r)| {
-                r.name
-                    .as_ref()
-                    .map(|n| n.to_lowercase().contains(&name_lower))
-                    .unwrap_or(false)
-            })
-            .collect();
-
-        assert_eq!(matching.len(), 2);
-        assert_eq!(matching[0].0, 0); // complaints-2023.csv
-        assert_eq!(matching[1].0, 2); // complaints-2024.csv
-    }
-
-    #[test]
-    fn test_resource_name_no_matches() {
-        // Create test resources
-        let resources = vec![
-            data_gov::ckan::models::Resource {
-                name: Some("data.csv".to_string()),
-                format: Some("CSV".to_string()),
-                url: Some("https://example.com/data.csv".to_string()),
-                ..Default::default()
-            },
-            data_gov::ckan::models::Resource {
-                name: Some("report.json".to_string()),
-                format: Some("JSON".to_string()),
-                url: Some("https://example.com/report.json".to_string()),
-                ..Default::default()
-            },
-        ];
-
-        // Test with no matches
-        let name_lower = "pdf".to_lowercase();
-        let matching: Vec<_> = resources
-            .iter()
-            .enumerate()
-            .filter(|(_, r)| {
-                r.name
-                    .as_ref()
-                    .map(|n| n.to_lowercase().contains(&name_lower))
-                    .unwrap_or(false)
-            })
-            .collect();
-
-        assert_eq!(matching.len(), 0);
-    }
-
-    #[test]
-    fn test_resource_selector_numeric_string_is_index() {
-        // "0" should be parsed as Index(0), not Name("0")
-        let result = ReplCommand::from_str("download dataset 0");
-        if let Ok(ReplCommand::Download { resource_selector, .. }) = result {
-            assert!(matches!(resource_selector, Some(ResourceSelector::Index(0))));
-        } else {
-            panic!("Expected Index, not Name");
-        }
-
-        // "999" should be parsed as Index(999), not Name("999")
-        let result = ReplCommand::from_str("download dataset 999");
-        if let Ok(ReplCommand::Download { resource_selector, .. }) = result {
-            assert!(matches!(resource_selector, Some(ResourceSelector::Index(999))));
-        } else {
-            panic!("Expected Index, not Name");
-        }
-    }
-
-    #[test]
-    fn test_resource_selector_mixed_string_is_name() {
-        // "0abc" cannot be parsed as usize, should be Name
-        let result = ReplCommand::from_str("download dataset 0abc");
-        if let Ok(ReplCommand::Download { resource_selector, .. }) = result {
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "0abc");
-            } else {
-                panic!("Expected Name(0abc)");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
-
-    #[test]
-    fn test_parse_command_args_simple() {
-        let args = parse_command_args("download dataset 0");
-        assert_eq!(args, vec!["download", "dataset", "0"]);
-    }
-
-    #[test]
-    fn test_parse_command_args_with_quotes() {
-        let args = parse_command_args("download dataset \"Comma Separated Values File\"");
-        assert_eq!(args, vec!["download", "dataset", "Comma Separated Values File"]);
-    }
-
-    #[test]
-    fn test_parse_command_args_multiple_spaces() {
-        let args = parse_command_args("search   climate    data");
-        assert_eq!(args, vec!["search", "climate", "data"]);
-    }
-
-    #[test]
-    fn test_parse_command_args_quotes_with_extra_spaces() {
-        let args = parse_command_args("download   dataset   \"Multi Word Name\"  ");
-        assert_eq!(args, vec!["download", "dataset", "Multi Word Name"]);
-    }
-
-    #[test]
-    fn test_parse_download_with_quoted_name() {
-        let result = ReplCommand::from_str("download my-dataset \"CSV File\"");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "my-dataset");
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "CSV File");
-            } else {
-                panic!("Expected Download command with Name(CSV File)");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
-
-    #[test]
-    fn test_parse_download_with_long_quoted_name() {
-        let result = ReplCommand::from_str("download dataset \"Comma Separated Values File\"");
-        assert!(result.is_ok());
-        
-        if let Ok(ReplCommand::Download { dataset_id, resource_selector }) = result {
-            assert_eq!(dataset_id, "dataset");
-            if let Some(ResourceSelector::Name(name)) = resource_selector {
-                assert_eq!(name, "Comma Separated Values File");
-            } else {
-                panic!("Expected Download command with quoted name");
-            }
-        } else {
-            panic!("Expected Download command");
-        }
-    }
 }
